+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Added support for `C50::C5.0()` classification tree models, including `decision_tree()` parsnip models fitted with the `"C5.0"` engine. Boosted (`trials > 1`) and rule-based (`rules = TRUE`) models are not supported. (#232)
 
+- Added support for `partykit::cforest()` random forest regression models, including `rand_forest()` parsnip models fitted with the `"partykit"` engine. Classification is not supported. (#232)
+
 - Added support for `LiblineaR::LiblineaR()` binary logistic regression models, including `logistic_reg()` parsnip models fitted with the `"LiblineaR"` engine. (#232)
 
 - Added support for `decision_tree()` parsnip models fitted with the `"rpart"` engine. (#232)